apps/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
cx_Oracle; platform_system == 'Windows' or platform_system == 'Linux' or platform_system == 'Darwin'
python-dotenv